run techproducts tests on cloud with two nodes
