+++ tests/Integration/SolrCloud/HttpTest.php
@@ -3,18 +3,18 @@
 namespace Solarium\Tests\Integration\SolrCloud;
 
 use Solarium\Core\Client\Adapter\Http;
-use Solarium\Tests\Integration\AbstractCollectionsTest;
+use Solarium\Tests\Integration\AbstractCloudTest;
 
 /**
  * @group integration
  * @group skip_for_solr_server
  * @coversNothing
  */
-class CollectionsHttpTest extends AbstractCollectionsTest
+class HttpTest extends AbstractCloudTest
 {
     public function setUp(): void
     {
         parent::setUp();
-        $this->client->setAdapter(new Http());
+        self::$client->setAdapter(new Http());
     }
 }
