+++ tests/Integration/SolrCloud/Psr18Test.php
@@ -2,13 +2,13 @@
 
 namespace Solarium\Tests\Integration\SolrCloud;
 
-use Solarium\Tests\Integration\AbstractCollectionsTest;
+use Solarium\Tests\Integration\AbstractCloudTest;
 
 /**
  * @group integration
  * @group skip_for_solr_server
  * @coversNothing
  */
-class CollectionsPsr18Test extends AbstractCollectionsTest
+class Psr18Test extends AbstractCloudTest
 {
 }
