+++ tests/Integration/SolrServer/Psr18Test.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Solarium\Tests\Integration\SolrServer;
+
+use Solarium\Tests\Integration\AbstractServerTest;
+
+/**
+ * @group integration
+ * @group skip_for_solr_cloud
+ * @coversNothing
+ */
+class Psr18Test extends AbstractServerTest
+{
+}
