--- tests/Integration/TechproductsAdapters/TechproductsPsr18Test.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace Solarium\Tests\Integration\TechproductsAdapters;
-
-use Solarium\Tests\Integration\AbstractCoreTest;
-
-/**
- * @group integration
- * @group skip_for_solr_cloud
- * @coversNothing
- */
-class TechproductsPsr18Test extends AbstractCoreTest
-{
-}
